fix(sdk): Remove stray call to node:Buffer
Update assertion.ts

fix(lib): Removes all refs to node:buffer I could find

- removes the polyfill from the web-test-runner, to make sure we aren't accidentally including it (at least in things that covers)
- updates all .spec and .test files, in turn
- renames a bunch of things called `buffer` to something else so I can search for this easier

Update cli.ts

+++ lib/tdf3/src/tdf.ts
@@ -838,7 +838,7 @@ export async function writeStream(cfg: EncryptConfiguration): Promise<DecoratedR
       throw new ConfigurationError(`Safe byte limit (${cfg.byteLimit}) exceeded`);
     }
     //new Uint8Array(chunk.buffer, chunk.byteOffset, chunk.byteLength);
-    crcCounter = unsigned(chunk as Uint8Array, crcCounter);
+    crcCounter = unsigned(chunk, crcCounter);
     fileByteCount += chunk.length;
   }
